datascope/requirements/development.txt
-r production.txt

django-debug-toolbar==1.9.1
django-extensions==1.9.0
Sphinx==1.3.1
sphinx-rtd-theme==0.1.8
pipdeptree==0.10.1
